+++ weblate/trans/util.py
@@ -332,6 +332,8 @@ def check_upload_method_permissions(user, translation, method: str):
             and user.has_perm("upload.perform", translation)
             and hasattr(translation.component.file_format_cls, "update_bilingual")
         )
+    if method == "add":
+        return user.has_perm("unit.add", translation)
     if method in ("translate", "fuzzy"):
         return user.has_perm("unit.edit", translation)
     if method == "suggest":